refactor: Ajuste na nomeação do domínio de grupo empresarial
