--- src/main/java/vagabounds/repositories/CompaniesGroupRepository.java
@@ -1,7 +1,0 @@
-package vagabounds.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import vagabounds.models.CompaniesGroup;
-
-public interface CompaniesGroupRepository extends JpaRepository<CompaniesGroup, Long> {
-}
